+++ src/app/(courses)/_types/index.ts
@@ -1,0 +1,12 @@
+import type { Student } from '@/types';
+
+export type Course = {
+  id: number
+  name: string
+  classroom: string
+  capacity: number
+  teacher: string
+  description?: string
+  active: boolean
+  students: Array<Student> 
+}
