+++ src/app/(courses)/course/[id]/_hooks/useCourse.ts
