+++ src/app/(courses)/course/create/_components/CreateCourseSuccess.tsx
