+++ src/app/(root)/favicon.ico
