+++ src/app/(root)/globals.css
