+++ src/app/(students)/_types/index.ts
@@ -1,0 +1,7 @@
+export type Student = {
+  id: number
+  name: string
+  email: string
+  age: number
+  gender: string
+}
