--- src/app/Courses/index.ts
@@ -1,3 +1,0 @@
-import Courses from './Courses';
-
-export default Courses;
