--- src/app/CreateStudentModal/index.ts
@@ -1,3 +1,0 @@
-import CreateStudentModal from './CreateStudentModal';
-
-export default CreateStudentModal;
