--- src/app/Providers.tsx
@@ -1,20 +1,0 @@
-'use client';
-
-import { ReactNode } from 'react';
-import { SnackbarProvider } from 'notistack';
-
-import { CssVarsProvider } from '@mui/joy/styles';
-
-import theme from '@/utils/theme';
-
-const Providers = ({ children }: { children: ReactNode }) => {
-  return (
-    <CssVarsProvider theme={theme}>
-      <SnackbarProvider>
-        {children}
-      </SnackbarProvider>
-    </CssVarsProvider>
-  );
-};
-
-export default Providers;
