--- src/app/course/create/CreateCourseDialog/index.ts
@@ -1,1 +1,0 @@
-export * from './CreateCourseDialog';
