--- src/app/course/create/page.tsx
@@ -1,11 +1,0 @@
-import Box from "@mui/joy/Box";
-
-import CreateCourseDialog from "@/app/course/create/CreateCourseDialog/CreateCourseDialog";
-
-export default function CourseCreate() {
-  return (
-    <Box display="flex" justifyContent="center" width="100%">
-      <CreateCourseDialog />
-    </Box>
-  );
-};
